Disable snapshot in publish Docker action
The snapshot option on elgohr/Publish-Docker-Github-Action creates
a unique tag using the date & time. This commit disables this option
as these "unique" images are not used in Inspektor Gadget, it's also
spamming our container registry with too many tags that are not relevant.
This commit also removes the duplicated push from the Makefile to
avoid pushing images tag based on the commit sha.

Signed-off-by: Mauricio Vásquez <mauricio@kinvolk.io>