gadgets: Implement Prometheus gadget
This commit implements a new Prometheus gadget that collects metrics
and exposes them to Prometheus.

The Prometheus gadget reuses the existing gadgets to collect the metrics.
It creates instances of them and registers a callback to get the events
they generate, then based on that it updates the metrics according to
the user configuration.

Check docs/design/001-prometheus.md to get more information about the
design of this feature

Implementation:

One key of this implementation is to be gadget neutral, it means there
is not any gadget-specific logic in this. It's generic and is able to
collect metrics from all gadgets. For this, we have to rely on the parser
(and hence columns) packages that abstract the types of the events that
each gadget provides.

- Counters:

The prometheus gadget registers a callback that is called each time the
gadget generates and event. For each event we need to:

1. Filter out the event: This is done by two different components,
well-known fields like namespace, pod and container are filtered by
the operators in eBPF. More specific fields are filtered by the already
existing filtering implementation in the parser.

2. Calculate attributes: Based on the labels configured by the user, we
need to get the attributes for the current event. This is implemented
by having a `AttrsGetter` method that returns a callback that generates
the attributes for a given event.

3. Increase the counter: Once we have the attributes, we can increase the
counter by using the otel meter.

- Gauges:

The prometheus gadget uses the snapshot gadgets to get a list of a given
resource. The filtering is configured as the counters, but in this case
the prometheus gadget itself aggregates entries as configured by the user.

TODO: (from easy to more difficult)
- other filtering operators
- float gauges
- histograms
- start / stop operations for kubectl-gadget case