go: Bump cilium/ebpf to 73acad5e3906
This upstream commit modifies how load from .kconfig are handled [1].
Before this commit, the register would not have bounds set:
; if (LINUX_KERNEL_VERSION < KERNEL_VERSION(5, 11, 0))
10: (18) r1 = 0xffffacb3808bc000      ; R1_w=map_value(off=0,ks=4,vs=4,imm=0)
12: (61) r1 = *(u32 *)(r1 +0)         ; R1_w=scalar(umax=4294967295,var_off=(0x0; 0xffffffff))
13: (67) r1 <<= 32                    ; R1_w=scalar(smax=9223372032559808512,umax=18446744069414584320,var_off=(0x0; 0xffffffff00000000),s32_min=0,s32_max=0,u32_max=0)
14: (c7) r1 s>>= 32                   ; R1_w=scalar(smin=-2147483648,smax=2147483647)
which leaded to some program not being loaded due to invalid access of ctx, for
example acessing ctx[1] while it does not exist for this kernel.
Indeed, without the bounds, the kernel could not remove branch in
is_branch*_taken().

With this commit, the bounds are correctly set:
; if (LINUX_KERNEL_VERSION < KERNEL_VERSION(5, 11, 0))
10: (18) r1 = 0xffffacb380988000      ; R1_w=map_value(off=0,ks=4,vs=4,imm=0)
12: (61) r1 = *(u32 *)(r1 +0)         ; R1_w=332561
13: (67) r1 <<= 32                    ; R1_w=1428338618925056
14: (c7) r1 s>>= 32                   ; R1_w=332561
So, the branch is removed and no bad access to ctx occurs.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
[1]: https://github.com/cilium/ebpf/commit/73acad5e3906