trace-network CR interface: use KubeIPResolver
KubeIPResolver is an operator that is automatically used when using the
gRPC interface. That's the case for the 'kubectl gadget trace network'
command.

But the 'kubectl gadget advise network-policy' still uses the CR
interface and not the gRPC interface. In this case, the KubeIPResolver
operator is not called automatically.

This patch calls the KubeIPResolver operator manually. Once the 'advise
network-policy' gadget uses the gRPC interface, we can remove this.