run: Support endpoint enrichment
This commit introduces a new endpoint type that can be used by the
events generated by the gadgets. The gadgets need to fill in the IP
address (IPv4 or IPv6), the version (4 or 6) and the port (in host byte
order) for each endpoint. Then the user space part automatically enriches
that with the KubeIPResolver operator.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>