socket enricher: make iterators optional
The socket enricher uses some bpf iterators but they are only available
on Linux >= 5.8. This patch moves the bpf iterators in a separate files
in order to have a fallback if it fails to load. Initial sockets and
delayed socket deletion will not work but at least we will still have
the main feature.

Signed-off-by: Alban Crequy <albancrequy@linux.microsoft.com>