socket enricher: support dual stack sockets
When gadget_socket_lookup() is trying to do a lookup for an IPv4 socket,
it also needs to do a second lookup for IPv6 and checks for
skc_ipv6only in case there is a dual stack socket.

After this patch, the socket enricher is able to handle the
kube-apiserver sockets:

```
$ go run -exec 'sudo' ./cmd/ig/... -r docker trace network -c minikube-docker
CONTAINER                    PID        TID        COMM            TYPE      PROTO PORT  REMOTE
minikube-docker              692650     692704     kubelet         OUTGOING  tcp   8443  192.168.58.2
minikube-docker              693180     693180     kube-apiserver  HOST      tcp   8443  192.168.58.2
```

An integration test is updated for this fix.

Signed-off-by: Alban Crequy <albancrequy@linux.microsoft.com>