docs: add examples how to run ig in a custom pod
The ig documentation included explanations how to run it with Docker and
how to run it with the "kubectl debug" command.

This patch adds another way to use ig in a custom pod. This shows an
example of Dockerfile and a Kubernetes daemon set yaml. This is useful
for third party software that wants to include ig.

Signed-off-by: Alban Crequy <albancrequy@linux.microsoft.com>