run: Support OCI images
This commit reworks the logic to support OCI images in the run command.
A requisite of this implementation is that OCI images should only be
pulled by the nodes, as it could happen that they're available in a
container registry that is not accessible by the client.

The main challenge of this implementation is to have the information
needed to create the parser, as it's done very soon in the workflow,
much before than running the gadget. In order to fix it, a new
GetGadgetInfo() method is added to the runtime, this is only relevant
for the run gadget.

Currenty the full BPF object bytes are sent to the client. Ideally only
the BTF information is needed, by right now it's not possible to
marshall a BTF type to JSON, and it's not easy to access the BTF raw
bytes for an eBPF object using the go-ebpf library.

The event decoding is still done in the client, hence, this commit
assumes that all events sent from the server have the same layout.
It's not a problem because:
- Inspektor Gadget only supports amd64 and arm64, hence endianess is not
an issue here.
- The event structure sent by the gadgets doesn't depend on the kernel
version. Also, the structure layout should be the same for different
architectures.
However we should revisit this soon to be sure the solution is generic
enough to support other architectures in the future.

This commit also organizes the different interfaces around by creating
a new RunGadgetDesc interface that is only implemented by the run gadget.
The motivation behind is that the run gadget will be the most important
(and perhaps the only) one we'll support in the future, then instead
of splitting pieces of it around, let's consolidate them in the same
place.

Co-authored-by: Mauricio Vásquez <mauriciov@microsoft.com>
Signed-off-by: Burak Ok <burakok@microsoft.com>
Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>