Create include folder with all header files exposed by IG
As we're moving more towards containerized gadgets, we need to prepare
everything needed to build gadgets outside the tree. The first step
is to consolidate all headers files used by gadgets in the same place.

This commit has a lot of changed files, but there is not any behaviour
change.

The only need thing, is the "make {install,remove}-headers" targets.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>