gadgets/run: Support eBPF parameters
Gadgets usually define different constants in the eBPF code to change
its behaviour. This commit implements the logic allow users setting
these flags from the CLI.

The parameters are gathered from a list provided in the gadget metadata
definition. Then the variable names are compared with the BTF
information and the corresponding variables will be replaced with the
user provided values.

Co-authored-by: Burak Ok <burakok@mircosoft.com>
Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>
Signed-off-by: Burak Ok <burakok@microsoft.com>