gadget-service: Add GetGadgetInfo() rpc
GetGadgetInfo is used by containerized gadgets to get data about the
gadget before running it. It's needed because in containerized gadgets
it's not possible to know features of the gadget before running it,
hence the logic is now split in GetGadgetInfo() and RunGadget().

GetGadgetInfoResponse has a generic bytes info for the time being, we
should declare specific types for that once the whole containerized
gadgets support settles a bit more.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>