gadgets/run: Implement GetGadgetInfo()
The main challenge of implementing support for OCI containerized gadgets
is to have the information needed to create the parser, as it's done
very soon in the workflow, much before than running the gadget. In order
to fix it, a new GetGadgetInfo() method is added to the runtime, this is
only relevant for the run gadget.

Currenty the full BPF object bytes are sent to the client. Ideally only
the BTF information is needed, by right now it's not possible to
marshall a BTF type to JSON, and it's not easy to access the BTF raw
bytes for an eBPF object using the go-ebpf library.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>