build: Automatically create or update metadata file
The metadata file contains information about the gadget. Creating
(and keeping) this file updated could be difficult, hence this commit
extends the build command to automatically create or update the file
according to the compiled ebpf program.

If the file doesn't exist, a new one is created with all default values,
if the file exists, ig tries to update it as best as it can, however
to avoid losing information introduced by the user, only new things are
added but nothing is removed.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>