run: Don't rely on map name
Relying on the map name to mark a map with a given purpose (e.g. print
maps) is not very flexible. It forces the user to have a specific naming
schema on their code, makes the map name longer, etc.

This commit introduces a GADGET_TRACE_MAP macro that marks a map as
tracer. The advantage is that the macro is defined by us and we could
change its implementation later on if needed.

This commit also renames all code from "print map" to "trace map" as
we already call that kind of gadgets "tracers" in our code base.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>