run: Add support for iterators
As in the tracers case, BTF is a key component as it's used to convert
binary data to a human readable format. Inspektor Gadget loads and
attaches all iterators programs of type task it found. When the gadget
is run, it reads the results of the iterator and uses the BTF
information provided above to format the data.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>