Experiment: Avoid decoding BTF on the client
This commit changes the logic to avoid the client having to decode BTF
information. The whole idea consists on having the "columns descriptors"
in the GadgetInfo structure. These descriptors are generated by the
server and used by the client to decode the data it receives.

The data is sent as a binary blob [][]byte. The first array is the bpf
event, the second 1 are fixed type fields and the other ones are used
for strings.

This commit also adds a BlobEvent struct that can be used by the server
to create other columns, the following commit shows how to use it.