run: Introduce gadget features concept
This commit introduces a new structure that describes the different
features that are provided by a containerized gadget. Those features
are filled by looking at the types, maps and programs present on the
bpf object.

This information will be used in next commits to make the handling
of containerized gadgets more flexible.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>