run: Send operator parameters in gadget info
Built-in gadgets get the parameters for the operators from the catalog.
This approach is, again, not valid for containerized gadgets as the
parameters could change from gadget to gadget. This commit extends
the GadgetInfo structure to have the operators parameter descriptor
and uses them in different places in the code.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>