run: Remove columns that aren't provided by the gadget
Columns are created from the Event definition in (pkg/gadgets/run/types/types.go),
however in the case of containerized gadgets that definition is generic
and the gadgets from provide all the fields described there.

This commit uses the gadgets features structure to remove the columns
that aren't available for a specific gadget.

TODO: Another approach to remove some of them is not to use
eventtypes.WithMountNsID & eventtypes.WithNetNsID, however the PR when
this commit is introduced is changing many things, so that is future
work.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>