operators: Import image layer operators in the OCI handler
Image layer operators are needed to handle the Gadget images. Before this
commit, users of the Golang API had to manually import the image layer
operators. It was a bit confusing as missing one operator could produce
misleading results (or not produce any result at all). By including these
operators on the OCI handler, we ensure that all the operators are available
to the users of the Golang API.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>