build: Use a cache on disk to speed up gadgets building
This commit uses the disk cache to make compilation of gadgets faster.
It could help in the following cases:
1. When editing the metadata of a gadget: There is net any need to
compile the whole gadget.
2. When editing the ebpf or wasm code in a gadget containing both: Only
the changed code should be compiled
3. When compiling the exact same version of the gadget: Useful on the
CI or when running "make -C gadgets/" to compile all of them.

The cache is implemented per layer (ebpf, wasm). The following things
are used to calculate the digest for the cache:
1. Source file creating the layer: program.bpf.c, wasm.go, etc.
2. Digest of the gadget builder image
3. Digest of all files under include/ or wasmapi/ when IG_SOURCE_PATH is
defined
4. Architecture

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>